model/requirements.txt
pytest==7.1.1
redis==4.1.4
lightgbm==2.0.3
pandas
numpy
boto3
lightgbm
Pillow
cloudpickle==3.0.0